int->float

+++ metric2/float64metric/float64metric.go
@@ -40,14 +40,14 @@ func (u UpDownCounter) Add(ctx context.Context, x float64, attrs ...attribute.Ke
 func (h Histogram) Record(ctx context.Context, x float64, attrs ...attribute.KeyValue) {
 }
 
-func (c Counter) Measure(x int64) batch.Measurement {
+func (c Counter) Measure(x float64) batch.Measurement {
 	return batch.Measurement{}
 }
 
-func (u UpDownCounter) Measure(x int64) batch.Measurement {
+func (u UpDownCounter) Measure(x float64) batch.Measurement {
 	return batch.Measurement{}
 }
 
-func (h Histogram) Measure(x int64) batch.Measurement {
+func (h Histogram) Measure(x float64) batch.Measurement {
 	return batch.Measurement{}
 }
